% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissmigDummyS.R
\name{kissmigDummyS}
\alias{kissmigDummyS}
\title{Get a simple suitability map}
\usage{
kissmigDummyS(mean, sd)
}
\arguments{
\item{mean}{temperature mean (degree celsius) of the suitability distribution}

\item{sd}{temperature standard deviation (degree celsius) of the  suitability distribution}
}
\description{
\command{kissmigDummyS} is a support function to generate simple suitability maps based on mean annual air temperature for example code.
}
\details{
\command{kissmigDummyS} returns a suitability map for a given extent based on mean annual temperature. It uses
data of WorldClim and calculates suitability as a normal distribution defined by \command{mean} and \command{sd}
of mean annual temperature. The density function is linearly rescaled to a maximum of 1, the maximum suitability
used in \command{kissmig}.
}
\references{
\url{https://www.worldclim.org/}
}
\seealso{
\code{\link{kissmig}}
}
