% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcl_bio1_europe.R
\docType{data}
\name{wcl_bio1_europe}
\alias{wcl_bio1_europe}
\title{Map of annual mean temperature covering Europe}
\format{
A \code{RasterLayer} object.
}
\source{
<https://geodata.ucdavis.edu/climate/worldclim/2_1/base/wc2.1_5m_bio.zip>
}
\usage{
wcl_bio1_europe
}
\description{
Map of the bioclimatic variable BIO1 (annual mean temperature in °C) covering Europe.
}
\details{
The map is based on WorldClim version 2.1 climate data for 1970-2000 and has
a spatial resolution of five minutes.
}
\references{
\url{https://www.worldclim.org}
}
\keyword{datasets}
