% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissmigAccess.R
\name{kissmigAccess}
\alias{kissmigAccess}
\title{Get accessiblity map from kissmig output}
\usage{
kissmigAccess(grd, rel=FALSE)
}
\arguments{
\item{grd}{a single RasterLayer of first occurrence generated by kissmig}

\item{rel}{if TRUE, kissmigAccess returns relative values with maximum 1, otherwise absolute interger values}
}
\description{
\command{kissmigAccess} calculates a accessibility map from a kissmig output of first occurrence (type='FOC'). These maps
allows the integration of limited migration in species distribution models and macroecological analyses.
}
\details{
\command{kissmig} maps of first occurrences show values of the first iteration step a raster cell was colonized. Early
colonized cells have low values, late colonized cells high values. These values are the opposite of accessibility,
which is high for early colonized, and low for late colonized cells. \command{kissmigAccess} simply calculates for each
cell the accessibility as the difference between the cell value and \command{max(grd)+1}. Cells which have never been
colonized remain unchanged (value 0).
}
\seealso{
\code{\link{kissmig}}
}
