% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissmigDummyS.R
\name{kissmigDummyS}
\alias{kissmigDummyS}
\title{Get a simple suitability map}
\usage{
kissmigDummyS(mean, sd)
}
\arguments{
\item{mean}{number of the mean temperature (degree celsius) of the suitability distribution.}

\item{sd}{number of the temperature standard deviation (degree celsius) of the  suitability distribution.}
}
\description{
\command{kissmigDummyS} is a support function to generate simple climate suitability maps
for example code.
}
\details{
\command{kissmigDummyS} returns a SpatRaster of a simple climate suitability map for Europe based on mean annual temperature. It uses
data of WorldClim and calculates suitability as a normal distribution defined by \code{mean} and \code{sd}
of mean annual temperature. The density function is linearly rescaled to a maximum of \code{1}, the maximum suitability
used in \code{\link{kissmig}}.
}
\references{
\url{https://www.worldclim.org/}
}
\seealso{
\code{\link{kissmig}}
}
