% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissmigOrigin.R
\name{kissmigOrigin}
\alias{kissmigOrigin}
\title{Define a geographic origin}
\usage{
kissmigOrigin(grd, x, y, size)
}
\arguments{
\item{grd}{a single RasterLayer as reference layer}

\item{x}{lower left x-coordinate of geographic origin}

\item{y}{lower left y-coordinate of geographic origin}

\item{size}{size of the quadratic origin}
}
\description{
\command{kissmigOrigin} is a support function to define the geographic origin for a \command{kissmig} call.
}
\details{
\command{kissmigOrigin} returns a rasterLayer characterized by the reference
\command{grd}. Cell values are set to zero, except for cells of the origin defined by
\command{extent(x, x+size, y, y+size)} which are set to one.
}
\seealso{
\code{\link{kissmig}}
}
